Support for Linux x86-64 architecture
* sort callbacks with NSInteger instead of int
* CGFloat is “double” in 64-bit, therefore it must be scanned with “%lg” instead of “%g”.
* NS(U)Integer is 64 bit wide when compiling for 64-bit. Therefore it must be formatted accordingly.
* NS(Un)Archiver must not treat pointers as four byte wide.
* CF(U)Integer is NS(U)Integer instead of int/unsigned
* objc_msgSendv implementation for x86-64
* NSInvocation
  - “setArgument:…”: check promotedSize against “sizeof(long)” instead of “4”
  - implemented “invokeWithTarget:” as documented by Apple
    → “invokeWithTarget:” sets the target first and then invokes
    → target changes
  - long != int → separate invocation

+++ Foundation/NSArray/NSMutableArray.m
@@ -381,7 +381,7 @@ static int _nsmutablearraycompareindices(const void* v1, const void* v2)
    [object release];
 }
 
-static int selectorCompare(id object1,id object2,void *userData){
+static NSInteger selectorCompare(id object1,id object2,void *userData){
    SEL selector=userData;
 
    return (NSComparisonResult)[object1 performSelector:selector withObject:object2];
